created custom element ts file
